cookbook/python-rag/requirements.txt
caskada>=1.0.0
numpy>=1.20.0
faiss-cpu>=1.7.0
openai>=1.0.0